% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/al_sample.R
\name{al_sample}
\alias{al_sample}
\title{Generate Age-Length sample}
\usage{
al_sample(N, ages, ps, lambda, strength_sd = 10)
}
\arguments{
\item{N}{total number of sample to be generated}

\item{ages}{vector of ages which distribs are based on}

\item{ps}{von bert param vector}

\item{lambda}{vector of proportion}

\item{strength_sd}{default = 10, greater the number smaller the sd}
}
\value{
tibble
}
\description{
Generate Age-Length sample
}
\examples{
\dontrun{
al_sample(100, ages = c(1, 2), ps = c(300, 0.6, 0), lambda = c(3, 1))}
}
