% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_optim.R
\name{estim_optim}
\alias{estim_optim}
\title{Estimate means and sds with multiple gaussian distribution to data}
\usage{
estim_optim(dat, means, sds)
}
\arguments{
\item{dat}{data with columns \code{c("Length", "Freq", "Freq_ratio")}}

\item{means}{means of each modes. number of means represents number of modes}

\item{sds}{sd of each distribution.
vector length must be the same as that of means}
}
\value{
tibble of parameters estimated
}
\description{
Estimate means and sds with multiple gaussian distribution to data
}
\examples{
\dontrun{
foodata <-
generate_sample(n = 100, ages = 3) \%>\%
freq_length() \%>\%
freq_ratio()

estim_optim(foodata, means = c(50, 80, 120), sds = c(10, 20, 30))
}
}
