% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_resids.R
\name{estim_resids}
\alias{estim_resids}
\title{Calculate sum of residual for specified means and sds}
\usage{
estim_resids(dat, means, sds)
}
\arguments{
\item{dat}{data with columns \code{c("Length", "Freq", "Freq_ratio")}}

\item{means}{means of each modes. number of means represents number of modes}

\item{sds}{sd of each distribution.
vector length must be the same as that of means}
}
\value{
sum of residuals
}
\description{
Calculate sum of residual for specified means and sds
}
\examples{
\dontrun{
foodata <-
generate_sample(n = 100, ages = 3) \%>\%
freq_length() \%>\%
freq_ratio()

estim_resids(dat = foodata, means = c(50, 80, 120), sds =c(10, 20, 30))}
}
