% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_em.R
\name{fit_em}
\alias{fit_em}
\title{Fit length data using EM algorithm}
\usage{
fit_em(dat, age)
}
\arguments{
\item{dat}{length vector}

\item{age}{number of modes of gaussian to be fitted}
}
\value{
Mclust result
}
\description{
Fit length data using EM algorithm
}
\examples{
\dontrun{
foo <- generate_sample(100, 3)
fit_em(foo, 3)}
}
