% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_em_range.R
\name{fit_em_range}
\alias{fit_em_range}
\title{fit_em for range of numbers of modes}
\usage{
fit_em_range(dat, Ages)
}
\arguments{
\item{dat}{length data to be fitted}

\item{Ages}{vector of numbers of modes}
}
\value{
tibble with nested mclust result
}
\description{
fit_em for range of numbers of modes
}
\examples{
\dontrun{
sampledata <- generate_sample(1000, 5)
fit_em_range(sampledata, 4)
fit_em_range(sampledata, c(1:10))}
}
