% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_em.R
\name{fit_range}
\alias{fit_range}
\title{Apply fit_em with various mode number to the data}
\usage{
fit_range(dat, ages = c(2:4))
}
\arguments{
\item{dat}{length vector}

\item{ages}{vector of number of modes to be applied}
}
\value{
listed Mclust result
}
\description{
Apply fit_em with various mode number to the data
}
\examples{
\dontrun{
foo <- generate_sample(100, 3)
fit_range(foo, ages = c(2:5))}
}
