% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{freq_ratio}
\alias{freq_ratio}
\title{Generate length frequency ratio}
\usage{
freq_ratio(dat)
}
\arguments{
\item{dat}{tibble with column = c(Length, Freq)}
}
\value{
tibble with Freq_ratio
}
\description{
Generate length frequency ratio
}
