% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample.R
\name{generate_sample}
\alias{generate_sample}
\title{Generate sample length data}
\usage{
generate_sample(n, ages, p = c(300, 0.2, 0))
}
\arguments{
\item{n}{Total number of sample}

\item{ages}{Number of ages, the maximum age}

\item{p}{Parameters of von Bertalanffy growth function
Linf: Asymptotic body length
K: growth coefficient
t0: Hypothetical age at size = 0}
}
\value{
Vector of generated length with sample size of n
}
\description{
Generate sample length data
}
\examples{
\dontrun{
foo <-generate_sample(n = 1000, ages = 7)
hist(foo)}
}
