% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{length_bin}
\alias{length_bin}
\title{Convert length into length bin}
\usage{
length_bin(dat, binwidth = 10)
}
\arguments{
\item{dat}{length data, vector}

\item{by}{bin size}
}
\value{
tibble with column \code{Length}
}
\description{
Convert length into length bin
}
