% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_resids.R
\name{resid2optim}
\alias{resid2optim}
\title{Calculate residuals with ps parameters}
\usage{
resid2optim(ps)
}
\arguments{
\item{ps}{parameters \code{c(means, sds)}}
}
\value{
sum of residuals
}
\description{
Calculate residuals with ps parameters
}
