% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/al_sample.R
\name{al_sample}
\alias{al_sample}
\title{Generate Age-Length sample with specified age and sd}
\usage{
al_sample(N, ages, ps, lambda, strength_sd = 10)
}
\arguments{
\item{N}{total number of sample to be produced}

\item{ages}{vector of ages}

\item{ps}{von bertalanffy parameters}

\item{lambda}{ratio of each gaussian. vector with length = that of ages}

\item{strength_sd}{sd = mean/strength_sd - ages*5, as greater the number, smaller the sd}
}
\value{
tibble
}
\description{
Generate Age-Length sample with specified age and sd
}
\examples{
\dontrun{
al_sample(100, c(1, 2, 3), ps = c(500, 0.5, -0.3), lambda = c(1, 2, 3))
}
}
