% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/al_sampler.R
\name{al_sample}
\alias{al_sample}
\title{Generate Age Length sample}
\usage{
al_sample(n, ages, ps, lambda, sdlevel = 10)
}
\arguments{
\item{n}{number of sample}

\item{ages}{age vector}

\item{ps}{vb parameter}

\item{lambda}{proportion of each age}

\item{sdlevel}{strength of sd, smaller the greater}
}
\value{
tibble
}
\description{
Generate Age Length sample
}
