% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_length.R
\name{freq_length}
\alias{freq_length}
\title{Create length frequency}
\usage{
freq_length(dat, binwidth = 10)
}
\arguments{
\item{dat}{length vector}

\item{binwidth}{width of length bin, 10 as standard}
}
\value{
tibble of length bin (1cm) and its frequency
}
\description{
Create length frequency
}
\examples{
\dontrun{
freq_length(
generate_sample(1000, 5), binwidth = 10)}
}
